pyproject.toml
[project]
name = "lan-party-stats"
version = "2.0.0"
description = "Discord bot that tracks game playtime and Spotify listening statistics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.2",
    "httpx>=0.26.0",
]

[project.scripts]
lan-party-bot = "main:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
]
